// src/model/pokemon.ts
export class Pokemon {

  abilities: Array<any>;
  base_experience: number;
  forms: Array<any>;
  game_indices: Array<any>;
  held_items: Array<any>;
  height: number;
  id: number;
  is_default: boolean;
  location_area_encounters: string;
  moves: Array<any>;
  name: string;
  order: number;
  past_types: Array<any>;
  species: {};
  sprites: {};
  stats: Array<any>;
  types: Array<any>;
  weight: number;
  description: {};
  gender: Array<any>;
  category: {};
  evolution: {};
  constructor(parameters) {

  }
}